solomontsega_portfolio_website/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv
